Use log_ceil, fn must have suffix attached

+++ src/utils.h
@@ -32,8 +32,8 @@ auto fupdate2(std::optional<T1> &x, const std::optional<T2> &x2, Fn fn) {
   return std::optional<decltype(fn(*x, *x2))>();
 }
 
-inline unsigned log2_ceil(unsigned count) {
-  unsigned bits = 0;
+inline uint64_t log2_ceil(uint64_t count) {
+  uint64_t bits = 0;
   while (count > (1 << bits)) {
     bits += 1;
   }
